WIP on setup.py (commit prior to big revert)
The situation with setuptools managing extension builds seems to
be that

 - setuptools' SWIG build is broken, requiring janky workarounds to
   sort of work (https://stackoverflow.com/questions/12491328),
   and by default it just spams SWIG's build artifacts in the source
   directory. Nice(!)
 - linking the setuptools-built module against shared libraries
   where the files are going to end up on the system in order to set
   R(UN)PATH correctly, which is impossible for a distributed module,
   because people might install Python modules anywhere. This cannot be
   worked around at runtime by setting LD_LIBRARY_PATH from the loader
   script as I had hoped, because the environment variables as far as
   the runtime linker are concerned are fixed as soon as the Python
   executable is launched. This is only a problem because OpenQL
   consists of multiple libraries referencing each other (lemon and
   libqasm), but let's be honest, that really shouldn't be causing
   this much drama.
 - linking the setuptools-built module against *static* libraries
   can't handle transitive dependencies. They'd have to be hardcoded
   into setup.py, but LEMON's dependency on glpk etc. is figured out by
   CMake at configure-time and thus cannot be hardcoded. Turns out that
   it's virtually impossible to pull this list out of CMake
   (https://gitlab.kitware.com/cmake/cmake/-/issues/12435), so this is a
   dead end. Also, it shouldn't even be possible to link static objects
   together into the shared object that Python needs, but that's what
   the original system was doing, so I stand corrected.

tl;dr setuptools' extension is horribly broken even for the simplest
case just on Linux, so I'll have to revert back to having CMake build
it.

+++ setup.py
@@ -7,12 +7,15 @@ from distutils.command.sdist import sdist as _sdist
 from distutils.command.build import build as _build
 from distutils.command.clean import clean as _clean
 from setuptools.command.egg_info import egg_info as _egg_info
+from setuptools.command.install import install as _install
+from setuptools.command.build_ext import build_ext as _build_ext
 import distutils.cmd
 import distutils.log
 from setuptools import setup, Extension, find_packages
 from wheel.bdist_wheel import bdist_wheel as _bdist_wheel
 import re
 
+src_dir = os.getcwd() + os.sep + 'src'          # source directory for C++
 python_dir = os.getcwd() + os.sep + 'python'    # source directory for Python
 pybuild_dir = os.getcwd() + os.sep + 'pybuild'  # python-specific build directory
 build_dir = pybuild_dir + os.sep + 'build'      # directory for setuptools to dump various files into
@@ -27,17 +30,12 @@ def get_version(verbose=0):
 
     matcher = re.compile('[\t ]*#define[\t ]+OPENQL_VERSION_STRING[\t ]+"(.*)"')
     version = None
-    try:
-        with open(os.path.join(srcDir, 'version.h'), 'r') as f:
-            for ln in f:
-                m = matcher.match(ln)
-                if m:
-                    version = m.group(1)
-                    break
-
-    except Exception as e:
-        print(e)
-        version = 'none'
+    with open(os.path.join(src_dir, 'version.h'), 'r') as f:
+        for ln in f:
+            m = matcher.match(ln)
+            if m:
+                version = m.group(1)
+                break
 
     if verbose:
         print('get_version: %s' % version)
@@ -51,8 +49,81 @@ def read(fname):
 class clean(_clean):
     def run(self):
         _clean.run(self)
-        shutil.rmtree(cbuild_dir)
-        shutil.rmtree(pybuild_dir)
+        if os.path.exists(cbuild_dir):
+            shutil.rmtree(cbuild_dir)
+        if os.path.exists(pybuild_dir):
+            shutil.rmtree(pybuild_dir)
+
+class build_ext(_build_ext):
+    def run(self):
+        from plumbum import local, FG
+
+        # Figure out how many parallel processes to build with.
+        if self.parallel:
+            nprocs = str(self.parallel)
+        else:
+            nprocs = os.environ.get('NPROCS', '1')
+
+        # Build the OpenQL C++ library using CMake and install it into
+        # prefix_dir.
+        if not os.path.exists(cbuild_dir):
+            os.makedirs(cbuild_dir)
+        with local.cwd(cbuild_dir):
+            (local['cmake']['..']
+                ['-DCMAKE_INSTALL_PREFIX=' + prefix_dir]
+                ['-DBUILD_SHARED_LIBS=YES'] # Transitive dependencies don't work with static libs
+                ['-DCMAKE_SKIP_RPATH=YES'] # see below (trigger warning!)
+            ) & FG
+            # We need to handle finding the shared objects manually at runtime.
+            # Normally, CMake will set the RPATH/RUNPATH key of the libraries
+            # it installs to the install directory, so it should find them even
+            # if they're not installed into a system directory and
+            # LD_LIBRARY_PATH isn't appropriately set by the user. However,
+            # CMake needs to know the exact absolute path where the files will
+            # end up in order to do this correctly. Since this is dynamically
+            # handled by Python and the instal; prefix directory certainly
+            # isn't correct (it gets deleted by pip after the install), all we
+            # can do is to have __init__.py set LD_LIBRARY_PATH to the correct
+            # value at runtime. Depending on Linux version, an RPATH key would
+            # override even LD_LIBRARY_PATH, so we have to disable the RPATH
+            # mechanism entirely during build/install.
+            #
+            # Note that the above RPATH nonsense only applies to Linux and Mac,
+            # and thankfully the same method should work for both. Windows is
+            # much more lenient in where it finds DLLs, checking the current
+            # directory and such by default, so it shouldn't need any of this
+            # stupidity.
+
+            # Do the build with the given number of parallel threads.
+            cmd = local['cmake']['--build']['.']
+            if nprocs != '1':
+                cmd = cmd['--parallel'][nprocs]
+            cmd & FG
+
+            # Do the install.
+            local['cmake']['--install']['.'] & FG
+
+        # Copy the user-written Python API files into module_dir for the
+        # parent build_ext to clobber with SWIG output and build to find the
+        # user modules AND swig-generated module in.
+        if os.path.exists(module_dir):
+            shutil.rmtree(module_dir)
+        shutil.copytree(python_dir, module_dir)
+
+        # Also copy the files CMake placed in lib/lib64 into the build
+        # directory managed by setuptools, so the shared objects/DLLs are
+        # added to the generated Python module.
+        dest_dir = self.build_lib + os.sep + 'openql'
+        if not os.path.exists(dest_dir):
+            os.makedirs(dest_dir)
+        src_dir = 'lib64' if os.path.exists(prefix_dir + os.sep + 'lib64') else 'lib'
+        src_dir = prefix_dir + os.sep + src_dir
+        for fname in os.listdir(src_dir):
+            if os.path.isfile(src_dir + os.sep + fname) and not os.path.islink(src_dir + os.sep + fname):
+                shutil.copyfile(src_dir + os.sep + fname, dest_dir + os.sep + fname)
+
+        # Let setuptools handle the SWIG build and final link.
+        _build_ext.run(self)
 
 class build(_build):
     def initialize_options(self):
@@ -60,43 +131,17 @@ class build(_build):
         self.build_base = os.path.relpath(build_dir)
 
     def run(self):
-        from plumbum import local, FG
-
-        # Build the OpenQL C++ library using CMake and install it into
-        # prefix_dir.
-        nprocs = os.environ.get('NPROCS', '1')
-        if not os.path.exists(cbuild_dir):
-            os.makedirs(cbuild_dir)
-        with local.cwd(cbuild_dir):
-            local['cmake']['..']['-DCMAKE_INSTALL_PREFIX=' + prefix_dir] & FG
-            local['cmake']['--build']['.']['--parallel'][nprocs] & FG
-            local['cmake']['--install']['.'] & FG
-
-        # Run SWIG on the installed library.
-        if not os.path.exists(swig_dir):
-            os.makedirs(swig_dir)
-        (local['swig']
-            ['-v']
-            ['-python']
-            ['-py3']
-            ['-castmode']
-            ['-modern']
-            ['-w511']
-            ['-c++']
-            ['-I' + prefix_dir + os.sep + 'include']
-            ['-outdir'][swig_dir]
-            ['-o'][swig_dir + os.sep + 'openql.cc']
-            [python_dir + os.sep + 'swig' + os.sep + 'openql.i']
-        ) & FG
-
-        # Copy the SWIG-generated file and the user-written Python API
-        # files into module_dir for setup to find.
-        if os.path.exists(module_dir):
-            shutil.rmtree(module_dir)
-        shutil.copytree(python_dir + os.sep + 'openql', module_dir)
-        shutil.copyfile(swig_dir + os.sep + 'openql.py', module_dir + os.sep + 'openql.py')
-
+        # Make sure the extension is built before the Python module is "built",
+        # otherwise SWIG's generated module isn't included.
+        # See https://stackoverflow.com/questions/12491328
+        self.run_command('build_ext')
         _build.run(self)
+
+class install(_install):
+    def run(self):
+        # See https://stackoverflow.com/questions/12491328
+        self.run_command('build_ext')
+        _install.run(self)
 
 class bdist(_bdist):
     def finalize_options(self):
@@ -136,8 +181,7 @@ setup(
     description='OpenQL Python Package',
     long_description=read('README.md'),
     long_description_content_type = 'text/markdown',
-    author='Nader Khammassi and Imran Ashraf',
-    author_email='nader.khammassi@gmail.com, iimran.aashraf@gmail.com',
+    author='QuTech, TU Delft',
     url='https://github.com/QE-Lab/OpenQL',
     license=read('LICENSE'),
 
@@ -158,24 +202,27 @@ setup(
     ],
 
     packages = ['openql'],
-    package_dir = {'openql': module_dir},
+    package_dir = {'openql': os.path.relpath(module_dir)},
 
     ext_modules = [
         Extension(
             'openql._openql',
-            [swig_dir + os.sep + '/openql.cc'],
+            [module_dir + os.sep + 'openql.i'],
             libraries = ['ql'],
             library_dirs = [prefix_dir + os.sep + 'lib', prefix_dir + os.sep + 'lib64'],
-            runtime_library_dirs = [prefix_dir + os.sep + 'lib', prefix_dir + os.sep + 'lib64'],
+            #runtime_library_dirs = [prefix_dir + os.sep + 'lib', prefix_dir + os.sep + 'lib64'],
             include_dirs = [prefix_dir + os.sep + 'include'],
-            extra_compile_args = ['-std=c++11']
+            extra_compile_args = ['-std=c++11'],
+            swig_opts = ['-v', '-py3', '-castmode', '-modern', '-w511', '-c++', '-I' + prefix_dir + os.sep + 'include']
         )
     ],
 
     cmdclass = {
         'bdist': bdist,
         'bdist_wheel': bdist_wheel,
+        'build_ext': build_ext,
         'build': build,
+        'install': install,
         'clean': clean,
         'egg_info': egg_info,
         'sdist': sdist,
